+++ mealplanner/app/src/main/java/comp3350/team10/presentation/MealDiaryLiveData.java
@@ -14,12 +14,12 @@ public class MealDiaryLiveData extends ViewModel {
 
     private MutableLiveData<Calendar> activityDate;             //The current date 
     private MutableLiveData<ArrayList<Edible>> mealsOnDate;  //The given meals on the current date
-    private MutableLiveData<Integer> goalCalories;              //The calorie goal for the current date
-    private MutableLiveData<Integer> consumedCalories;          //The calories consumed for the current date
-    private MutableLiveData<Integer> exerciseCalories;          //The calories burnt during exercise for the current date
-    private MutableLiveData<Integer> netCalories;               //The net calorie calculation for a given date (consumed - burnt)
-    private MutableLiveData<Integer> progressBar;               //The progress bar for the current date
-    private MutableLiveData<Integer> progressExcess;            //The excess progress bar for the current date 
+    private MutableLiveData<Double> goalCalories;              //The calorie goal for the current date
+    private MutableLiveData<Double> consumedCalories;          //The calories consumed for the current date
+    private MutableLiveData<Double> exerciseCalories;          //The calories burnt during exercise for the current date
+    private MutableLiveData<Double> netCalories;               //The net calorie calculation for a given date (consumed - burnt)
+    private MutableLiveData<Double> progressBar;               //The progress bar for the current date
+    private MutableLiveData<Double> progressExcess;            //The excess progress bar for the current date
                                                                 //(different than progress bar)
     public MutableLiveData<Calendar> getActivityDate() {
         if (activityDate == null) {
@@ -36,49 +36,49 @@ public class MealDiaryLiveData extends ViewModel {
         return mealsOnDate;
     }
 
-    public MutableLiveData<Integer> getGoalCalories() {
+    public MutableLiveData<Double> getGoalCalories() {
         if (goalCalories == null) {
-            goalCalories = new MutableLiveData<Integer>(1500);
+            goalCalories = new MutableLiveData<Double>(1500.0);
         }
 
         return goalCalories;
     }
 
-    public MutableLiveData<Integer> getConsumedCalories() {
+    public MutableLiveData<Double> getConsumedCalories() {
         if (consumedCalories == null) {
-            consumedCalories = new MutableLiveData<Integer>(1000);
+            consumedCalories = new MutableLiveData<Double>(1000.0);
         }
 
         return consumedCalories;
     }
 
-    public MutableLiveData<Integer> getExerciselCalories() {
+    public MutableLiveData<Double> getExerciselCalories() {
         if (exerciseCalories == null) {
-            exerciseCalories = new MutableLiveData<Integer>(200);
+            exerciseCalories = new MutableLiveData<Double>(200.0);
         }
 
         return exerciseCalories;
     }
 
-    public MutableLiveData<Integer> getNetCalories() {
+    public MutableLiveData<Double> getNetCalories() {
         if (netCalories == null) {
-            netCalories = new MutableLiveData<Integer>(69);
+            netCalories = new MutableLiveData<Double>(69.0);
         }
 
         return netCalories;
     }
 
-    public MutableLiveData<Integer> getProgressBar() {
+    public MutableLiveData<Double> getProgressBar() {
         if (progressBar == null) {
-            progressBar = new MutableLiveData<Integer>(0);
+            progressBar = new MutableLiveData<Double>(0.0);
         }
 
         return progressBar;
     }
 
-    public MutableLiveData<Integer> getProgressExcess() {
+    public MutableLiveData<Double> getProgressExcess() {
         if (progressExcess == null) {
-            progressExcess = new MutableLiveData<Integer>(0);
+            progressExcess = new MutableLiveData<Double>(0.0);
         }
 
         return progressExcess;
